move&refactor: Map 관련 API 분리
- Map 관련 API는 독서노트가 아니라 책지도 화면에서 주고 받는 API이므로 url의 계층 표시를 명확히 하기 위해 분리함
